+++ app/scripts/local-storage.ts
@@ -1,6 +1,4 @@
 /** @format */
-
-import omit from "lodash/omit"
 import { SearchParams } from "@/urlparams"
 import { toJson } from "angular"
 
@@ -13,27 +11,6 @@ export const localStorageGet = <K extends keyof LocalStorage>(key: K): LocalStor
 /** Write object to local storage. To delete, use native `localStorage.removeItem(key)`. */
 export const localStorageSet = <K extends keyof LocalStorage>(key: K, value: LocalStorage[K]): void =>
     localStorage.setItem(key, toJson(value))
-
-/**
- * Convert old jStorage data to native localStorage.
- */
-export function convertJstorage(): void {
-    const jstorageData = localStorage.getItem("jStorage")
-    if (!jstorageData) return
-
-    try {
-        const json = jstorageData.replace(/^"(.*)"$/, "$1")
-        const data = omit(JSON.parse(json), "__jstorage_meta") as LocalStorage
-        console.log("Converting local storage:", data)
-        Object.keys(data).forEach((key: keyof LocalStorage) => {
-            // Write to new storage if empty
-            if (!localStorageGet(key)) localStorageSet(key, data[key])
-        })
-        localStorage.removeItem("jStorage")
-    } catch (error) {
-        console.error("Error converting old jStorage data:", error)
-    }
-}
 
 /** Defines local storage objects used by this app. */
 export type LocalStorage = {